requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-vision==3.1.4
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
